move utils and constants
