+++ src/seed_data/factories/Course.factory.ts
@@ -1,33 +1,11 @@
 import { define } from 'typeorm-seeding';
 import { Course } from '../../entity/Course';
+import { subjects, getRandomIntInclusive, toUpperCase } from '../utils'
 
 import * as faker from 'faker';
 
 faker.seed(123);
 
-// TODO move to consts
-const subjects = ['PHMD', 'PHSC', 'AB', 'CD', 'BIOL', 'CHEM'];
-// =================================================
-
-
-// TODO move to utils
-function getRandomIntInclusive(min, max) {
-    min = Math.ceil(min);
-    max = Math.floor(max);
-    return Math.floor(Math.random() * (max - min + 1) + min); //The maximum is inclusive and the minimum is inclusive
-}
-
-function toUpperCase(value) {
-    const words = value.split(' ');
-    var i: number;
-    for (i = 0; i < words.length; i++) {
-        words[i] = words[i][0].toUpperCase() + words[i].substr(1);
-    }
-
-    return words.join(' ');
-}
-
-// =================================================
 let courseNumbers = [];
 
 function generateCourseNumber() {
