+++ src/seed_data/utils/constants.ts
@@ -1,0 +1,7 @@
+export const subjects = ['PHMD', 'PHSC', 'AB', 'CD', 'BIOL', 'CHEM'];
+export const semesters = ['FL', 'SP', 'SU', 'S1', 'S2'];
+export const statuses = ['ENROLLED', 'LEAVE', "DROP_BACK", 'COOP', 'GRADUATED'];
+export const entryTypes = ['EA', 'DE'];
+export const startYears = [2018, 2017, 2016];
+export const endYears = [2020, 2021, 2022];
+export const years = [2021, 2020, 2019, 2018, 2017, 2016];
