+++ src/seed_data/utils/utilFunctions.ts
@@ -1,0 +1,51 @@
+import * as faker from 'faker';
+import { LetterGrade } from '../../entity/Enums';
+
+export function getRandomIntInclusive(min: number, max: number) {
+    min = Math.ceil(min);
+    max = Math.floor(max);
+    return Math.floor(Math.random() * (max - min + 1) + min); //The maximum is inclusive and the minimum is inclusive
+}
+
+export function toUpperCase(value: String) {
+    const words = value.split(' ');
+    var i: number;
+    for (i = 0; i < words.length; i++) {
+        words[i] = words[i][0].toUpperCase() + words[i].substr(1);
+    }
+
+    return words.join(' ');
+}
+
+export function chooseFrom(options: any[]) {
+    const x = faker.random.number(options.length - 1);
+    return options[x];
+}
+
+export function chooseLetterGrade(percentage: number) {
+    if (percentage >= 93) {
+        return LetterGrade.A;
+    } else if (percentage >= 90) {
+        return LetterGrade.A_MINUS;
+    } else if (percentage >= 87) {
+        return LetterGrade.B_PLUS;
+    } else if (percentage >= 83) {
+        return LetterGrade.B;
+    } else if (percentage >= 80) {
+        return LetterGrade.B_MINUS;
+    } else if (percentage >= 77) {
+        return LetterGrade.C_PLUS;
+    } else if (percentage >= 73) {
+        return LetterGrade.C;
+    } else if (percentage >= 70) {
+        return LetterGrade.C_MINUS;
+    } else if (percentage >= 67) {
+        return LetterGrade.D_PLUS;
+    } else if (percentage >= 65) {
+        return LetterGrade.D;
+    } else if (percentage >= 63) {
+        return LetterGrade.D_MINUS;
+    } else {
+        return LetterGrade.F;
+    }
+}
